Add new MissionReporter class
