+++ version_b/main.rb
@@ -1,3 +1,4 @@
+require_relative 'mission_reporter'
 require_relative 'mission_control'
 require_relative 'mission_plan'
 require_relative 'mission'
